cleanup_repo: Remove hardcoded repo path
The /repo path was hardcoded in the cleanup_repo.py script. Even if
there is an ansible variable (dci_retrieve_component_local_repo) it
was not honored.
The script now uses the complete repo_path (local_repo and product
name) as a parameter for the script.
Also remove the unused python library StringIO.

Change-Id: I10753f51c38eb7453b17ecdd281cbe886f047f08