Always run sync_repo tasks
When a task from sync_repo.yml (mainly reposync) fails for any reason,
the next time the role is executed for the same component the
sync_repo.yml tasks won't be executed as the .repo file did not change.
This means that when reposync fails (network error, …) the agent enters
a state where it won't try to download the component again and will try
to run a deployment against inconsistent data.
Always running tasks from sync_repo.yml ensures that the agent will try
to get the repo in a consistent state each time the agent executes.

Change-Id: I8464fdea169bf2bdcdf03ed28e67c56a82176eac